Fix: FireBase 초기화 클래스 추가 및 홈 디테일 이동 추가
