+++ app/build.gradle
@@ -1,7 +1,5 @@
 plugins {
     alias(libs.plugins.android.application)
-    // Google 서비스 Gradle 플러그인 추가
-    id 'com.google.gms.google-services'
 }
 
 android {
@@ -14,7 +12,6 @@ android {
         targetSdk 34
         versionCode 1
         versionName "1.0"
-
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
