+++ app/src/main/res/layout/fragment_contest_recruitment_detail.xml
@@ -84,7 +84,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_marginStart="8dp"
-                    android:text="뒤로가기"
+                    android:text="내 참여 공모전 목록"
                     android:textColor="@android:color/black"
                     android:textStyle="bold"
                     android:textSize="16sp" />
