+++ app/src/main/res/layout/fragment_home.xml
@@ -2,7 +2,6 @@
     android:layout_height="match_parent"
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:background="@android:color/white"
-    android:padding="24dp"
     >
 
     <LinearLayout
